requirements.txt
# Bear GitOps Payment Service

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
pydantic>=2.5.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
httpx>=0.25.0
